child-safe-search-ecosystem/requirements.txt
# === AI Providers ===
anthropic>=0.49.0
google-genai>=1.0.0
openai>=1.70.0

# === Search APIs ===
tavily-python>=0.5.0
httpx>=0.27.0          # Serper (no official SDK — raw HTTP)

# === Web Scraping ===
beautifulsoup4>=4.12.0
lxml>=5.3.0
playwright>=1.48.0

# === API Framework ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# === Config & Validation ===
python-dotenv>=1.0.0
pydantic>=2.9.0

# === Reliability ===
tenacity>=9.0.0
